Crate e1000

Modules

test_e1000_driver

Structs

E1000Registers
struct to hold mapping of registers

E1000Nic
struct representing an e1000 network interface card.

RegistersRx
struct to hold registers related to one receive queue

RegistersTx
struct to hold registers related to one transmit queue

Constants

E1000_DEV

INTEL_VEND

Functions

get_e1000_nic
Returns a reference to the E1000Nic wrapped in a MutexIrqSafe, if it exists and has been initialized.