Struct e1000::E1000Nic

pub struct E1000Nic { /* private fields */ }

Struct representing an e1000 network interface card.

Implementations

impl E1000Nic

Functions that setup the NIC struct and handle the sending and receiving of packets.

pub fn init(
e1000_pci_dev: &PciDevice
) -> Result<&'static MutexIrqSafe<E1000Nic>, &'static str>

Initializes the new E1000 network interface card that is connected as the given PciDevice.

pub fn spoof_mac(&mut self, spoofed_mac_addr: [u8; 6])