Struct e1000::E1000Nic

pub struct E1000Nic { /* private fields */ }

Struct representing an e1000 network interface card.

Implementations

impl E1000Nic

Functions that setup the NIC struct and handle the sending and receiving of packets.

pub fn init(
e1000_pci_dev: &PciDevice
) -> Result<&'static MutexIrqSafe<E1000Nic>, &'static str>

Initializes the new E1000 network interface card that is connected as the given PciDevice.

enable_interrupts must be called after the NIC has been registered with the net subsystem.

pub fn init_interrupts(
&mut self,
interface: Arc<NetworkInterface>
) -> Result<(), &'static str>

Initializes the interrupt handler and enables interrupts for this E1000 NIC.

The provided interface must be the network interface associated with this E1000 NIC. This interface will be polled in a deferred task upon an interrupt being triggered for a received packet.

pub fn spoof_mac(&mut self, spoofed_mac_addr: [u8; 6])